@@ -1,22 +1,19 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
 using ServiceStack;
 
-namespace MyApp
+[assembly: HostingStartup(typeof(MyApp.ConfigureUi))]
+
+namespace MyApp;
+
+public class ConfigureUi : IHostingStartup
 {
-    public class ConfigureUi : IConfigureAppHost
-    {
-        public void Configure(IAppHost appHost)
-        {
+    public void Configure(IWebHostBuilder builder) => builder
+        .ConfigureAppHost(appHost => {
             // if wwwroot/ is empty, build Client App with 'npm run build'
-            var svgDir = appHost.RootDirectory.GetDirectory("/svg") ?? appHost.ContentRootDirectory.GetDirectory("/src/static/svg"); 
+            var svgDir = appHost.RootDirectory.GetDirectory("/assets/svg");
             if (svgDir != null)
             {
                 Svg.Load(svgDir);
             }
-        }
-    }
+            Svg.CssFillColor["svg-icons"] = "#FF3E00";
+        });
 }
-
